GetLines math changes
1) I think perhaps the bounds on the if/else are reversed from what they
should be; if we're setting x values, shouldn't that mean the line is
closer to horizontal, meaning that t <45 or >135?

2) If I understand how the math was set up, I think the current code
will output values for X and Y as if they are still centered on the
origin. This will be an easy fix later if I am correct. Note that if the
if/else conditional is flipped then it won't be right on the first try.