Potential Angle Fix
I noticed that a lot of the lines we were getting in our current output
image from the Hough voting were at 45 degree angles. That made me think
about the angles we were using to detect things and remembered that it
was originally coded for a square image. (I believe this came from the
open source code we started with.) I'm hoping this will make at least
some of the lines look less random and will potentially address some of
the lines trailing off.